+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `loading` prop for `Image` to allow loading priority.
+
 ## [2.89.0] - 2023-12-21
 
 ### Added